// packages/hardhat-ignition/packages/core/src/internal/execution/types/execution-state.ts
export enum ExecutionStatus {
  STARTED = "STARATED",
  TIMEOUT = "TIMEOUT",
  SUCCESS = "SUCCESS",
  HELD = "HELD",
  FAILED = "FAILED",
}

/**
 * The different kinds of execution states.
 */
export enum ExecutionSateType {
  DEPLOYMENT_EXECUTION_STATE = "DEPLOYMENT_EXECUTION_STATE",
  CALL_EXECUTION_STATE = "CALL_EXECUTION_STATE",
  STATIC_CALL_EXECUTION_STATE = "STATIC_CALL_EXECUTION_STATE",
  ENCODE_FUNCTION_CALL_EXECUTION_STATE = "ENCODE_FUNCTION_CALL_EXECUTION_STATE",
  CONTRACT_AT_EXECUTION_STATE = "CONTRACT_AT_EXECUTION_STATE",
  READ_EVENT_ARGUMENT_EXECUTION_STATE = "READ_EVENT_ARGUMENT_EXECUTION_STATE",
  SEND_DATA_EXECUTION_STATE = "SEND_DATA_EXECUTION_STATE",
}

/**
 * The Strategy configuration that the user has provided in the run
 * the future is started. Used to reconcile the future on subsequent runs.
 */
export type ConcreteExecutionConfig = Record<string, number | string>;

/**
 * The base interface for all execution states.
 *
 * Its id must match the id of the future that it belongs to.
 */
interface BaseExecutionState<
  ExecutionStateT extends ExecutionSateType,
  FutureTypeT extends FutureType
> {
  id: string;
  type: ExecutionStateT;
  futureType: FutureTypeT;
  strategy: string; // For example, "basic" | "create2". This needs to be string if we want custom ones.
  strategyConfig: ConcreteExecutionConfig;
  status: ExecutionStatus;
  dependencies: Set<string>;
}
// ...
export interface SendDataExecutionState
  extends BaseExecutionState<
    ExecutionSateType.SEND_DATA_EXECUTION_STATE,
    FutureType.SEND_DATA
  > {
  to: string;
  data: string;
  value: bigint;
  from: string;
  networkInteractions: NetworkInteraction[];
  result?: SendDataExecutionResult;
}